+++ src/test/java/org/ojalgo/optimisation/ModelFileMPS.java
@@ -47,9 +47,6 @@ public interface ModelFileMPS {
     static void assertValues(final ExpressionsBasedModel model, final BigDecimal expMinVal, final BigDecimal expMaxVal, final Map<String, BigDecimal> solution,
             final NumberContext precision) {
 
-        //        model.options.debug(LinearSolver.class);
-        //        model.options.mip_defer = 0.0;
-
         TestUtils.assertTrue(model.validate());
 
         if (expMinVal != null) {
@@ -94,6 +91,13 @@ public interface ModelFileMPS {
 
         ExpressionsBasedModel model = ModelFileMPS.makeModel(dataset, modelName, relax);
 
+        // model.options.debug(IntegerSolver.class);
+        // model.options.debug(ConvexSolver.class);
+        // model.options.debug(LinearSolver.class);
+        // model.options.progress(IntegerSolver.class);
+        // model.options.validate = false;
+        // model.options.mip_defer = 0.0;
+
         BigDecimal expMinVal = expMinValString != null ? new BigDecimal(expMinValString) : null;
         BigDecimal expMaxVal = expMaxValString != null ? new BigDecimal(expMaxValString) : null;
 
@@ -121,11 +125,6 @@ public interface ModelFileMPS {
 
             model.options.time_suffice = 5L * CalendarDateUnit.MINUTE.toDurationInMillis();
             model.options.time_abort = 15L * CalendarDateUnit.MINUTE.toDurationInMillis();
-
-            // model.options.debug(IntegerSolver.class);
-            // model.options.progress(IntegerSolver.class);
-            // model.options.debug(LinearSolver.class);
-            // model.options.validate = false;
 
             TestUtils.assertTrue(model.validate());